tools/etl/requirements.txt
# Core geospatial stack
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# HTTP + env
requests>=2.31.0
python-dotenv>=1.0.0

# Supabase
supabase>=2.8.0